Add GET, PUT, POST and REST end points for tags.
The URL is of the form:

    http://api.{ninja-site}/rest/v1/sites/{site-id}/tags/{tag-id}

the content type must be application/json.

Tags are named JSON objects associated with a site and stored in the cloud.

Tags created with PUT and POST. PUT always replaces the latest version
without archival, POST will archive the current version and create a new version.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>